refactor(group): add exit group function and add user id params

+++ src/group/group.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 
 import { CreateGroupDto } from './dto/create-group.dto';
 import { UpdateGroupDto } from './dto/update-group.dto';
@@ -18,6 +18,7 @@ export class GroupService {
         data: {
           description,
           name,
+          admin_id: userId,
           created_by_id: userId,
           group_membership: {
             createMany: {
@@ -25,9 +26,7 @@ export class GroupService {
             },
           },
         },
-        include: {
-          group_membership: true,
-        },
+        include: { group_membership: true },
       }),
     ]);
 
@@ -36,15 +35,13 @@ export class GroupService {
 
   async findAll() {
     const groups = await this.prismaService.group.findMany({
-      include: {
-        group_membership: true,
-      },
+      include: { group_membership: true },
     });
     return groups;
   }
 
   async updateGroup(groupId: string, updateGroupDto: UpdateGroupDto) {
-    const { description, name } = updateGroupDto;
+    const { description, name, new_admin } = updateGroupDto;
 
     const updatedGroup = await this.prismaService.$transaction([
       this.prismaService.group.update({
@@ -52,6 +49,7 @@ export class GroupService {
         data: {
           description,
           name,
+          admin_id: new_admin,
         },
       }),
     ]);
@@ -59,8 +57,18 @@ export class GroupService {
     return updatedGroup;
   }
 
-  async addMembers(addMemberDto: AddMemberDto) {
+  async addMembers(addMemberDto: AddMemberDto, user_id: string) {
     const { group_id, members } = addMemberDto;
+
+    const { admin_id } = await this.prismaService.group.findFirst({
+      where: { id: group_id },
+    });
+
+    if (admin_id !== user_id)
+      throw new UnauthorizedException(
+        'You were not permitted to add member this group',
+      );
+
     const updatedMember = await this.prismaService.$transaction(
       members.map((member) =>
         this.prismaService.groupMembership.create({
@@ -75,8 +83,17 @@ export class GroupService {
     return updatedMember;
   }
 
-  async deleteMembers(deleteMemberDto: DeleteMemberDto) {
+  async deleteMembers(deleteMemberDto: DeleteMemberDto, user_id: string) {
     const { group_id, memberships } = deleteMemberDto;
+
+    const { admin_id } = await this.prismaService.group.findFirst({
+      where: { id: group_id },
+    });
+
+    if (admin_id !== user_id)
+      throw new UnauthorizedException(
+        'You were not permitted to add member this group',
+      );
 
     const updatedMember = await this.prismaService.$transaction(
       memberships.map((membership) =>
@@ -92,7 +109,29 @@ export class GroupService {
     return updatedMember;
   }
 
-  async remove(groupId: string) {
+  async exitGroup(group_id: string, user_id: string) {
+    await this.prismaService.groupMembership.delete({
+      where: {
+        user_id_group_id: {
+          group_id,
+          user_id,
+        },
+      },
+    });
+  }
+
+  async remove(groupId: string, user_id: string) {
+    const { admin_id } = await this.prismaService.group.findFirst({
+      where: {
+        id: groupId,
+      },
+    });
+
+    if (admin_id !== user_id)
+      throw new UnauthorizedException(
+        'You were not permitted to delete this group',
+      );
+
     await this.prismaService.$transaction([
       this.prismaService.groupMembership.deleteMany({
         where: { group_id: groupId },
